Reflective sphere in engine tests
Managed to create reflection of a scene in an object in engine tests programm.
TODO: HEAVY refactoring, maybe consider making new engine, current engine too hard to use on more complex things like this one with reflections. Also all the classes for buffer object are too much and are puzzling even for me