Engine1 v.0.2.6
Added small class that renders point and spot lights as cubes in the game world. This can be very useful for debugging lights and other stuff.
All cubes have the same size which is specified in class constructor and can be changed later
Cubes are colored the same as light source they represent
Light cubes dont affect shadow maps and mirrors